# docs/swagger.yaml
basePath: /
definitions:
...
  handlers.CreateProductRequest:
    properties:
      category:
        type: string
      description:
        type: string
      images:
        items:
          type: string
        type: array
      name:
        type: string
      price:
        type: number
      stock:
        type: integer
    type: object
...
  handlers.UpdateProductRequest:
    properties:
      category:
        type: string
      description:
        type: string
      id:
        type: integer
      images:
        items:
          type: string
        type: array
      name:
        type: string
      price:
        type: number
      status:
        type: string
      stock:
        type: integer
    type: object
...
  models.HotProduct:
    properties:
      productId:
        type: integer
      productName:
        type: string
      salesAmount:
        type: number
      salesCount:
        type: integer
    type: object
...
  models.Product:
    description: 商品信息模型
    properties:
      category:
        type: string
      createdAt:
        type: string
      description:
        type: string
      id:
        type: integer
      images:
        items:
          type: string
        type: array
      name:
        type: string
      price:
        type: number
      status:
        description: on_sale, off_sale, deleted
        type: string
      stock:
        type: integer
      updatedAt:
        type: string
    type: object
  models.SalesStat:
    properties:
      date:
        type: string
      orderCount:
        type: integer
      totalSales:
        type: number
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: PetShop 宠物商店 API 文档 - 提供宠物管理、购物车、订单管理、用户管理等功能
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: PetShop API
  version: "1.0"
paths:
...
  /api/admin/product:
    delete:
      consumes:
      - application/json
      description: 将商品标记为已删除（软删除）
      parameters:
      - description: 商品ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功消息
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 请求参数错误
          schema:
            type: string
        "401":
          description: 未授权
          schema:
            type: string
        "404":
          description: 商品不存在
          schema:
            type: string
        "409":
          description: 商品有待处理订单
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 删除商品
      tags:
      - 商品管理
    get:
      consumes:
      - application/json
      description: 根据ID获取商品详细信息
      parameters:
      - description: 商品ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 商品详情
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: 请求参数错误
          schema:
            type: string
        "401":
          description: 未授权
          schema:
            type: string
        "404":
          description: 商品不存在
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 获取商品详情
      tags:
      - 商品管理
    put:
      consumes:
      - application/json
      description: 更新现有商品信息
      parameters:
      - description: 商品更新请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新后的商品
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: 请求参数错误
          schema:
            type: string
        "401":
          description: 未授权
          schema:
            type: string
        "404":
          description: 商品不存在
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 更新商品
      tags:
      - 商品管理
  /api/admin/products:
    get:
      consumes:
      - application/json
      description: 获取所有商品列表（不包括已删除的）
      produces:
      - application/json
      responses:
        "200":
          description: 商品列表
          schema:
            items:
              $ref: '#/definitions/models.Product'
            type: array
        "401":
          description: 未授权
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 获取商品列表
      tags:
      - 商品管理
    post:
      consumes:
      - application/json
      description: 创建新商品
      parameters:
      - description: 商品创建请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 创建成功的商品
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: 请求参数错误
          schema:
            type: string
        "401":
          description: 未授权
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 创建商品
      tags:
      - 商品管理
  /api/admin/stats/hot-products:
    get:
      consumes:
      - application/json
      description: 获取销量最高的商品列表
      parameters:
      - default: 10
        description: 返回数量限制
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 热销商品列表
          schema:
            items:
              $ref: '#/definitions/models.HotProduct'
            type: array
        "401":
          description: 未授权
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 获取热销商品
      tags:
      - 销售统计
  /api/admin/stats/sales:
    get:
      consumes:
      - application/json
      description: 获取销售统计数据，支持日、周、月维度
      parameters:
      - default: day
        description: 统计周期 (day:最近7天, week:最近4周, month:最近6月)
        in: query
        name: period
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 销售统计数据
          schema:
            items:
              $ref: '#/definitions/models.SalesStat'
            type: array
        "401":
          description: 未授权
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 获取销售统计
      tags:
      - 销售统计
...
securityDefinitions:
  APITokenAuth:
    description: 使用 API Token 进行认证，用于 Open API 访问
    in: header
    name: X-API-Token
    type: apiKey
  BearerAuth:
    description: 使用 JWT Token 进行认证，格式：Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"